Separate Bio from About section
* Add links to header
* Use parallax
* Add band image
