+++ src/pages/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Link from 'gatsby-link'
 import Hero from '../components/Hero'
+import About from '../components/About'
 import Bio from '../components/Bio'
 import Listen from '../components/Listen'
 import PreOrder from '../components/PreOrder'
@@ -9,6 +10,7 @@ import Footer from '../components/Footer'
 const IndexPage = () => (
   <div>
     <Hero />
+    <About />
     <Bio />
     <Listen />
     <PreOrder />
